[FTR] ListView Widget within the Application Widget
Feature Development

The overall list of ingredients for all recipes is currently populating the list.  So overall
modifications have been performed that has been the only way I have been able to get results.

- 1. initializing the data in the Constructor for the view seemed to not aid in anything.
- 2. Moved persistant data query to the `onDataSetChanged` method.  I am not familar with
     when or why this method gets called within the stack.  Originally thought it should be
     similar to that of a `RecyclerViewer` widget `Adapter` data notification change but this
     was not the case.
- 3. Full cursor context is stored but dont see why not an already fetched set of data would
     not work either.
- 4. Debugging the populating remote view adapter is quite difficult as no information is
     easily provided in the terminal.  Don't know why this is.

Related Issues
- #28
- #29
