+++ app/src/main/res/layout/baking_widget.xml
@@ -3,7 +3,7 @@
     android:id="@+id/ll_widget"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="#09C"
+    android:background="#FFF"
     android:padding="@dimen/widget_margin"
     android:orientation="vertical">
 
@@ -11,29 +11,22 @@
         android:orientation="vertical"
         android:layout_width="match_parent"
         android:layout_height="wrap_content">
-
         <ListView
             android:id="@+id/lv_widget_ingredient_list"
-            android:layout_width="wrap_content"
-            android:layout_height="0dp"
-            android:layout_weight="9" >
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content" >
         </ListView>
-        <TextView
-            android:id="@+id/empty_view"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/default_ingreident_text"/>
+    </LinearLayout>
 
+    <FrameLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
         <TextView
             android:id="@+id/tv_widget_recipe_name"
             android:layout_width="wrap_content"
-            android:layout_height="0dp"
-            android:layout_weight="1"
+            android:layout_height="wrap_content"
             android:textSize="12sp"
             android:layout_gravity="center"
             android:textStyle="bold|italic" />
-
-    </LinearLayout>
-
-
+    </FrameLayout>
 </LinearLayout>
